requirements.txt
dash==2.16.1
numpy==1.26.2
pandas==2.2.1
plotly==5.20.0
pyspark==3.5.1
requests==2.31.0
scipy==1.11.4
sympy==1.12
word2number==1.1
pyzmq~=26.1.1
tqdm~=4.66.5
pytest~=8.3.2
dill~=0.3.8
redis~=5.1.0
msgspec~=0.18.6